backend/requirements.txt
fastapi
uvicorn
pydantic-ai
langgraph
pydantic
sqlalchemy
pgvector
boto3
python-dotenv
httpx
